movies that fall out of the ordering (e.g. new ones) are displayed in a separate list
I also added a cascade on the movie_id foreign key in the
movie-positions table. This way I can actually delete a movie record, as
long as I have removed it from the order beforehand, i.e. there is no
movie_position record that references it anymore.

Movies that fall out of the order dictated by movie_positions are listed
separately in an Inbox list. This list is meant to contain movies that I
haven't had a chance to order yet (e.g. new ones). This is also the only
list from which I can actually delete items, so I removed the delete
button from the other list.

The two lists are drag-drop connected. This means that I can take a
movie off the ordered list by dragging it in the Inbox and vice versa.